chore: removes unused hooks from usePipelineRuns.ts
